add a syscall() wrapper that falls back to fiddle
This is pretty gross. Ruby 1.9.2 has no clean way (AFAICS) to
detect whether or not Kernel.syscall works until it throws an
exception. This module tries it and falls back to fiddle which
is default in 1.9.2. libffi support is planned & stubbed as
a final fallback that should work anywhere, possibly even in
jruby.