checkpoint, still broken
It's still hanging on Kernel.spawn inside the container.

Strace is showing it selecting on fd 4 which is a pipe; this shouldn't be happening
until the spawn returns. Using plain fork/exec does the same thing. Ruby is doing something
weird, so I might take a pass at directly calling fork/exec with FFI or look at the ruby
source to see why it's trying to read on my pipes (close_on_exec?).